Eric Sunshine resolved circular inclusion of event headers. Also fixed a few other compilation errors and warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21759 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/cstool/pen.cpp
@@ -304,7 +304,7 @@ csPen::Write(iFont *font, uint x1, uint y1, char *text)
 
   pos += tt;
 
-  g2d->Write(font, pos.x, pos.y, the_color, -1, text);  
+  g2d->Write(font, (int)pos.x, (int)pos.y, the_color, -1, text);  
 }
 
 void
